small changes to width of header
found a bug! The top of my page does not show up in browser!
You can see everything except the header in firefox. Some of the intro
section is cut off in edge. I guess I sized it to look OK in firefox,
and it is visibile when you're in inspector mode, but it is messed up :(